Light & Dark mode themes refactored NOTE: 1. applyTheme() function in theme.js file
