+++ frontend/JS/main.js
@@ -1,3 +1,5 @@
+import applyTheme from "./theme.js";
+
 const sampleProducts = [
   {
     id: 1,
@@ -86,32 +88,7 @@ document.addEventListener("DOMContentLoaded", async () => {
     }
   }
 
-  // ---------------------------
-  // 🔹 Dark mode toggle
-  // ---------------------------
-
-  const themeBtn = document.querySelector(".theme-btn");
-  const body = document.body;
-
-  if (themeBtn) {
-    // Apply saved theme
-    const savedTheme = localStorage.getItem("theme");
-    if (savedTheme === "dark") {
-      body.classList.add("dark-mode");
-      themeBtn.textContent = "☀️";
-    }
-
-    themeBtn.addEventListener("click", () => {
-      body.classList.toggle("dark-mode");
-      const isDark = body.classList.contains("dark-mode");
-
-      themeBtn.textContent = isDark ? "☀️" : "🌙";
-      themeBtn.classList.add("rotate");
-      setTimeout(() => themeBtn.classList.remove("rotate"), 500);
-
-      localStorage.setItem("theme", isDark ? "dark" : "light");
-    });
-  }
+  applyTheme();
 
   const accountBtn = document.getElementById("account-btn");
   if (accountBtn) {
