+++ frontend/cart.html
@@ -5,7 +5,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>سبد خرید - فروشگاه قاب آنلاین</title>
   <link rel="stylesheet" href="CSS/style.css">
-  <script src="JS/cart.js" defer></script>
+  <script type="module" src="JS/cart.js" defer></script>
 </head>
 <body>
