+++ frontend/profile.html
@@ -6,7 +6,7 @@
   <title>حساب من</title>
 
   <link rel="stylesheet" href="CSS/style.css">
-  <script src="JS/profile.js" defer></script>
+  <script type="module" src="JS/profile.js" defer></script>
 </head>
 <body>
