+++ frontend/signup.html
@@ -5,7 +5,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>ثبت نام | فروشگاه قاب آنلاین</title>
   <link rel="stylesheet" href="css/style.css">
-  <script src="js/auth.js" defer></script>
+  <script type="module" src="js/auth.js" defer></script>
 </head>
 <body class="signup-page">
